Properly source `ripgrep` configuration and ignore brew lock file
Found out this was an issue when trying to use `C-sc` and saw that it
was showing no files. It wasn't just hidden files that was showing, but
it was not following hidden files by default. This was because the
ripgrep configuration was not properly sourced, so it has now been added
to the `.zshrc`.